pyproject.toml
[project]
name = "typosniffer"
version = "0.1.0"
description = "A tool to generate/find potential typosquatting attempts agains a domain"
authors = [
    {name = "PierKnight",email = "pier.altimari@libero.it"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13, <3.14"
dependencies = [
    "dnstwist (>=20250130,<20250131)",
    "requests (>=2.32.5,<3.0.0)",
    "typeguard (>=4.4.4,<5.0.0)",
    "rich-click (>=1.8.9,<2.0.0)",
    "dnspython (>=2.7.0,<3.0.0)",
    "textdistance[extras] (>=4.6.3,<5.0.0)",
    "tldextract (>=5.3.0,<6.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "whoisit (>=3.1.1,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pydantic[email] (>=2.11.9,<3.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "selenium (>=4.35.0,<5.0.0)",
    "imagehash (>=4.3.2,<5.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "torchvision (>=0.23.0,<0.24.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "numpy (>=2.3.3,<3.0.0)",
]

[tool.poetry]
packages = [{include = "typosniffer", from = "src"}]

[project.scripts]
typosniffer = "typosniffer.cli.cli:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
